pyproject.toml
[project]
name = "py001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
]
